eryn priors with extra burnin phase
Added in eryn priors and included an extra step to allow for burnin.

The format for eryn (final .hdf5) is a little different from the .hdf5 file
that is produced using emcee. This makes it incompatible with the package _arviz_.

Need to resolve this, or change plotting utilities if we wish to change to eryn.